+++ tests/test_labelled_fields.py
@@ -3,8 +3,7 @@ import pytest
 
 from phillydb.labelled_fields import label_city_owned_properties
 
+
 def test_label_city_owned_properties():
     with pytest.raises(ValueError):
         label_city_owned_properties(pd.DataFrame())
-
-